prakruti-backend/requirements.txt
# Requirements for PRAKRUTI backend with AI chatbot
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.24.3
pydantic-settings==2.1.0
pydantic==2.5.0
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0

# AI API dependencies
openai==1.3.7
google-generativeai==0.3.2
